.history/requirements_20251106164123.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
openai==1.12.0
pinecone-client==3.0.3
python-dotenv==1.0.0
pymongo==4.6.1
python-docx==1.1.0
tiktoken==0.5.2
pydantic==2.5.3
pydantic-settings==2.1.0
langchain==0.1.6
langchain-openai==0.0.5
langchain-community==0.0.20




